backend/hackaton/requirements.txt
asgiref==3.8.1
astroid==3.3.5
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.9
Django==5.1.3
django-cors-headers==4.6.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mccabe==0.7.0
openapi-codec==1.3.2
packaging==24.2
platformdirs==4.3.6
psycopg2==2.9.10
PyJWT==2.10.0
pylint==3.3.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.2
tomlkit==0.13.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
